Environment manager dynamically adds zones to map
