+++ src/states/zone1/zone1.config.js
@@ -8,3 +8,8 @@ export const TREES = new Set([
     { x: 2250, y: window.innerHeight - 40, scale: 1 },
     { x: 4000, y: window.innerHeight - 40, scale: 1.75 }
 ]);
+
+export const CONFIG = {
+  WORLD_WIDTH,
+  TREES
+};
