+++ src/states/zone1/zone1.js
@@ -1,9 +1,9 @@
-import { EnvironmentManager } from './environment.manager';
+import { EnvironmentManager } from '../../environment/environment.manager';
 import { PlayerManager } from '../../player/player.manager';
 import { LayerManager } from '../../layerManager/layerManager';
 import { EnemyManager } from '../../enemyManager/enemyManager';
 import { GrassSprite } from '../../environment/grass.sprite';
-import { WORLD_WIDTH } from './zone1.config';
+import { CONFIG } from './zone1.config';
 
 export class Zone1 extends Phaser.State {
 
@@ -38,7 +38,7 @@ export class Zone1 extends Phaser.State {
     this.player = new PlayerManager({ game: this.game });
     this.game.enemyManager.addZombie( { player: this.player.sprite } );
     this.ground = new GrassSprite({ game: this.game });
-    this.environment = new EnvironmentManager({ game: this.game });
+    this.environment = new EnvironmentManager({ game: this.game, config: CONFIG });
     /**
      * Phaser.World.setBounds(x, y, width, height )
      * x — Top left most corner of the world..
@@ -46,7 +46,7 @@ export class Zone1 extends Phaser.State {
      * width — New width of the game world in pixels.
      * height — New height of the game world in pixels.
      */
-    this.game.world.setBounds(0, 0, WORLD_WIDTH, window.innerHeight);
+    this.game.world.setBounds(0, 0, CONFIG.WORLD_WIDTH, window.innerHeight);
   }
 
 }
